Convert set::Peer into fragments
Part of #2
