+++ src/set/mod.rs
@@ -4,3 +4,5 @@ mod device;
 pub use device::{Device, WgDeviceF};
 mod peer;
 pub use peer::{Peer, WgPeerF};
+mod peer_fragment;
+pub(crate) use peer_fragment::PeerFragment;
